requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
scikit-learn>=1.3.0
seaborn>=0.12.0